Refactor ledger tests to use getTransactions()
The Ledger class is intended to be primarily used with
the getTransactions() by clients, with arguments specifying
restrictions while functions like getAllTransactionsInTimeRange(),
etc. are used internally as an implementation detail in enforcing
those restrictions.

As such, tests should be concerned with making sure that the API,
getTransactions(), works as intended, rather than the implementation
details realizing this.